Fully funcational scripts just needs finalization & sudoers file fix

+++ post-provision.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+
+## Configure shell variables
 OUT=/etc/dite_pki
 CA=$OUT/ca
 CERTS=$OUT/certs
@@ -9,8 +11,11 @@ CA_KEY=$CA/Dite_CA/private/Dite_CA.key
 CA_ROOT_PATH=$CA/Dite_CA
 CA_SIGNING_PATH=$CA/Intermediate_CA
 
+#### CA passwords
+export SIGNING_PASSPHRASE='654321'
+export ROOT_PASSPHRASE='123456'
 
-
+## Create the required base directories
 mkdir -pv /etc/dite_pki
 mkdir -pv /var/www/html/private
 mkdir -pv /var/www/html/private/notes
@@ -20,14 +25,16 @@ mkdir -pv $CA_SIGNING_PATH
 mkdir -pv $CA_ROOT_PATH
 mkdir -pv $CA/Dite_CA/private
 
-# configure environment so that the required tools are installed on the system
-##This should only be the base requirements
-##Ensure that the CA certificate is configured accordingly
+## This should only be the base requirements
+# #Ensure that the CA certificate is configured accordingly
 
-echo "Disable SELinux"
+
+## configure disabled SELinux
+echo "Disable SELinux"ex
 sed -i 's/permissive/disabled/g' /etc/selinux/config
 sestatus
 
+## Proceed to prepare package installtion
 yum remove -y docker \
                   docker-client \
                   docker-client-latest \
@@ -37,45 +44,80 @@ yum remove -y docker \
                   docker-logrotate \
                   docker-engine
 
+# Install yum repository
 yum install -y yum-utils
 
+## Install Docker CE on the server
 yum-config-manager \
     --enable \
     --add-repo \
-    https://download.docker.com/linux/centos/docker-ce.repo
-#
-dnf update -y
-#
-dnf install -y epel-release python3-cryptography python3-pexpect.noarch python36.x86_64 python3-firewall python3-pip  epel-release python3-cryptography docker-ce python3-pexpect.noarch docker-ce-cli containerd.io open-vm-tools expect #python3-docker.noarch
-#
-echo "Install packages required for OpenEMR"
-dnf install -y python36.x86_64 python3-firewall python3-pip python3-cryptography  php-mbstring mariadb-server mariadb php-xml.x86_64 python3-PyMySQL.noarch python3-libselinux.x86_64 epel-release php-json.x86_64 python3-pexpect.noarch httpd php php-mysqlnd mod_ssl
-#
-cp -rf /srv/hosts/pki/etc/* $CONFIG
+    /srv/hosts/config/docker-ce.repo
+
+#dnf update -y
+
+dnf install -y epel-release docker-ce docker-ce-cli containerd.io open-vm-tools expect
 
 systemctl start docker
-cp -rf /srv/hosts/config/daemon.json /etc/docker/daemon.json
-systemctl restart docker
+
+# Ensure that the service is started
+# systemctl restart docker
 systemctl enable docker
-##
-##
-#
-##
+
+# Add the docker user
+useradd -G docker -u 1040 -s /bin/bash -m dockerdev
+
+## ~Install http packages and configuration as required
+echo "Install packages required for OpenEMR operations"
+dnf install -y php-mbstring mariadb-server mariadb php-xml.x86_64 python3-PyMySQL.noarch python3-libselinux.x86_64 php-json.x86_64 httpd php php-mysqlnd mod_ssl
+
+\cp -rf /srv/hosts/config/httppassword /etc/httpd/htpasswd
+
+\cp -rf /srv/hosts/config/apache2.config /etc/httpd/conf.d/ssl.conf
+
+\cp -rf /srv/hosts/config/docker.service /etc/systemd/system/multi-user.target.wants/docker.service
+
+systemctl daemon-reload
+\cp /srv/hosts/web/* /var/www/html/
+
+## Download the OpenEMR packaging system
+
+if [[ ! -e /var/www/html/openemr-5_0_1_3 ]]; then
+    curl -L http://192.168.56.1/openemr-5_0_1_3.tar.gz -o /tmp/openemr-5_0_1_3.tar.gz
+    tar xvf /tmp/openemr-5_0_1_3.tar.gz -C /var/www/html/
+fi
+
+OpenEMRFiles=('sites/default/sqlconf.php' 'interface/modules/zend_modules/config/application.config.php' 'sites/default/documents' 'sites/default/edi' 'sites/default/era' 'sites/default/letter_templates' 'gacl/admin/templates_c' 'interface/main/calendar/modules/PostCalendar/pntemplates/compiled' 'interface/main/calendar/modules/PostCalendar/pntemplates/cache')
+
+systemctl start mariadb
+
+for i in "${OpenEMRFiles[@]}"; do
+    chmod 0666 /var/www/html/openemr-5_0_1_3/"$i"
+done
+
+### Provision database as intended
+mysql < /srv/hosts/config/openemr_doc.sql
+
+## Copy base config files and HTML images
+\cp -rf /srv/hosts/pki/etc/* $CONFIG
+
+\cp -rf /srv/hosts/web/index.php /var/www/html/
+\cp -rf /srv/hosts/images/background.jpg /var/www/html/
+\cp -rf /srv/hosts/web/index.html /var/www/html/private/
+\cp -rf /srv/hosts/web/style.css /var/www/html/private/
+
 ### 1. Create Root CA
-##
+
 if [[ ! -e $CA_ROOT_PATH/db/Dite_CA.db ]]; then
     echo "Creating database files required for CA"
     mkdir -p $CA_ROOT_PATH/private $CA_ROOT_PATH/db $CRL $CERTS
     chmod 700 $CA_ROOT_PATH/private
 
-    cp /dev/null $CA_ROOT_PATH/db/Dite_CA.db
-    cp /dev/null $CA_ROOT_PATH/db/Dite_CA.db.attr
+    \cp /dev/null $CA_ROOT_PATH/db/Dite_CA.db
+    \cp /dev/null $CA_ROOT_PATH/db/Dite_CA.db.attr
     echo 01 > $CA_ROOT_PATH/db/Dite_CA.crt.srl
     echo 01 > $CA_ROOT_PATH/db/Dite_CA.crl.srl
 fi
-#
-export ROOT_PASSPHRASE='123456'
-#
+
 if [[ ! -e $CA/Dite_CA.crt ]]; then
 
     openssl req -new \
@@ -92,20 +134,21 @@ if [[ ! -e $CA/Dite_CA.crt ]]; then
         -out $CA/Dite_CA.crt \
         -extensions ext_ca_root
 fi
-#
-#
+
 ## 2. Create Signing CA
 
-mkdir -p $CA_SIGNING_PATH/private $CA_SIGNING_PATH/db crl certs
-chmod 700 $CA_SIGNING_PATH/private
+if [[ ! -e $CA_SIGNING_PATH/db/Intermediate_CA.db ]]; then
 
-cp /dev/null $CA_SIGNING_PATH/db/Intermediate_CA.db
-cp /dev/null $CA_SIGNING_PATH/db/Intermediate_CA.db.attr
-echo 01 > $CA_SIGNING_PATH/db/Intermediate_CA.crt.srl
-echo 01 > $CA_SIGNING_PATH/db/Intermediate_CA.crl.srl
-#
-export SIGNING_PASSPHRASE='654321'
-#
+    echo "Creating database files required for CA"
+    mkdir -p $CA_SIGNING_PATH/private $CA_SIGNING_PATH/db crl certs
+    chmod 700 $CA_SIGNING_PATH/private
+
+    \cp /dev/null $CA_SIGNING_PATH/db/Intermediate_CA.db
+    \cp /dev/null $CA_SIGNING_PATH/db/Intermediate_CA.db.attr
+    echo 01 > $CA_SIGNING_PATH/db/Intermediate_CA.crt.srl
+    echo 01 > $CA_SIGNING_PATH/db/Intermediate_CA.crl.srl
+fi
+
 if [[ ! -e $CA_SIGNING_PATH/Intermediate_CA.crt ]]; then
 
     openssl req -new \
@@ -175,102 +218,52 @@ fi
 #
 ### 3.1. e-mail
 #
-export FRED_PASSPHRASE='a12345'
+export DOCKER_PASSPHRASE='a12345'
 #
-if [[ ! -e $CERTS/fred.p12 ]]; then
+if [[ ! -e $CERTS/dockerdev.p12 ]]; then
 
   openssl req -new \
-      -passout env:FRED_PASSPHRASE \
-      -subj="/DC=re/O=Dite Inc/CN=Fred Flintstone/emailAddress=fred@dite.re/" \
+      -passout env:DOCKER_PASSPHRASE \
+      -subj="/DC=re/O=Dite Inc/CN=Docker Development/emailAddress=dockerdev@dite.re/" \
       -config $CONFIG/csr-email.conf \
-      -out $CERTS/fred.csr \
-      -keyout $CERTS/fred.key
+      -out $CERTS/dockerdev.csr \
+      -keyout $CERTS/dockerdev.key
 
   openssl ca \
       -batch \
       -passin env:SIGNING_PASSPHRASE \
       -config $CONFIG/ca-signing.conf \
-      -in $CERTS/fred.csr \
-      -out $CERTS/fred.crt \
+      -in $CERTS/dockerdev.csr \
+      -out $CERTS/dockerdev.crt \
       -extensions ext_email
 
   openssl pkcs12 -export \
-      -passin env:FRED_PASSPHRASE \
-      -passout env:FRED_PASSPHRASE \
-      -name "Fred Flintstone" \
-      -inkey $CERTS/fred.key \
-      -in $CERTS/fred.crt \
-      -out $CERTS/fred.p12
+      -passin env:DOCKER_PASSPHRASE \
+      -passout env:DOCKER_PASSPHRASE \
+      -name "Docker Development" \
+      -inkey $CERTS/dockerdev.key \
+      -in $CERTS/dockerdev.crt \
+      -out $CERTS/dockerdev.p12
 
 fi
-#
-## Create PEM bundle
-#cat $CERTS/fred.key \
-#    $CERTS/fred.crt \
-#    > $CERTS/fred.pem
-#
 
-#
-#create_certificate() {
-#
-#    if [[ ! -e $CERTS/$1.crt ]]; then
-#        subjectAltName=$2 \
-#        openssl req -new \
-#            -config $CONFIG/csr-server.conf \
-#            -out $CERTS/$1.csr \
-#            -keyout $CERTS/$1.key \
-#            -subj="$DEFAULT_DIST_NAME /CN=$1"
-#
-#        openssl ca \
-#            -batch \
-#            -config $CONFIG/ca-signing.conf \
-#            -in $CERTS/$2.csr \
-#            -out $CERTS/$2.crt \
-#            -passin env:SIGNING_PASSPHRASE \
-#            -extensions ext_server
-#    fi
-#}
-#
-#if [[ ! -e $CERTS/acreage.local.crt ]]; then
-#
-#    subjectAltName=DNS:acreage.dite.local\
-#    openssl req -new \
-#        -config $CONFIG/csr-server.conf \
-#        -out $CERTS/acreage.org.csr \
-#        -keyout $CERTS/acreage.org.key \
-#        -subj="/DC=re/O=Dite Inc/CN=acreage.dite.local"
-#
-#
-#    openssl ca \
-#        -batch \
-#        -config $CONFIG/ca-signing.conf \
-#        -in $CERTS/acreage.org.csr \
-#        -out $CERTS/acreage.org.crt \
-#        -passin env:SIGNING_PASSPHRASE \
-#        -extensions ext_server
-#
-#fi
-#
-## 3.6 Create CRL
-#
+
 openssl ca -gencrl \
     -batch \
     -config $CONFIG/ca-signing.conf \
     -out $CRL/ca-signing.crl \
     -passin env:SIGNING_PASSPHRASE
 #
-cp /srv/hosts/web/* /var/www/html/
-#
-if [[ ! -e /var/www/html/openemr-5_0_1_3 ]]; then
-    curl -L http://172.16.48.181/openemr-5_0_1_3.tar.gz -o /tmp/openemr-5_0_1_3.tar.gz
-    tar xvf /tmp/openemr-5_0_1_3.tar.gz -C /var/www/html/
-fi
 
-#cp /srv/hosts/config/apache2.config /etc/httpd/conf.d/ssl.conf
+#\cp -rf /srv/hosts/config/apache2.config /etc/httpd/conf.d/ssl.conf
 #
-cp $CA/ca-signing-chain.pem /usr/share/pki/ca-trust-source/anchors/
+\cp $CA/ca-signing-chain.pem /usr/share/pki/ca-trust-source/anchors/
 #
 /usr/bin/update-ca-trust
+
+\cp -rf $CA/ca-signing-chain.pem /var/www/html/private
+\cp -rf $CERTS/dockerdev.p12 /var/www/html/private
+
 #
 ### Copy CA certificate to trust store on the system
 #
@@ -278,8 +271,7 @@ cp $CA/ca-signing-chain.pem /usr/share/pki/ca-trust-source/anchors/
 #
 ###Generate SSH public key
 
-
-cp -rf /srv/hosts/pki/id_rsa* /etc/dite_pki/
+\cp -rf /srv/hosts/pki/id_rsa* /etc/dite_pki/
 
 /usr/bin/expect <<EOD | grep ssh-rsa > /etc/dite_pki/ca-ssh.pub
 spawn /usr/bin/ssh-keygen -f $CA_SIGNING_PATH/private/Intermediate_CA.key -y
@@ -288,7 +280,6 @@ send "654321\n"
 expect eof
 EOD
 
-
 /usr/bin/expect <<EOD
 spawn /usr/bin/ssh-keygen -s $CA_SIGNING_PATH/private/Intermediate_CA.key -I 'edcbb' -z 0003 -n root /etc/dite_pki/id_rsa.pub
 expect "Enter passphrase:"
@@ -296,19 +287,23 @@ send "654321\n"
 expect eof
 EOD
 
-cp -rf /srv/hosts/pki/id_rsa* /etc/dite_pki/
+\cp -rf /srv/hosts/pki/id_rsa* /etc/dite_pki/
 
 systemctl restart httpd
 systemctl enable httpd
+\cp -rf /srv/hosts/config/daemon.json /etc/docker/daemon.json
+systemctl restart docker
+systemctl enable docker
+
+\cp -rf /etc/dite_pki/ca-ssh.pub /srv/hosts/pki/ca-ssh.pub
+\cp -rf $CA/ca-signing-chain.pem /srv/hosts/pki/ca-signing-chain.pem
+systemctl restart sshd
 
 
-#bersekr bleach eclipse cicada
 
-#
-#/usr/bin/ssh-keygen -s $CA_SIGNING_PATH/private/Dite_Intermediate_CA.key -I 'edcbb' -z '0003 -n dockerdev $CA/dockerdev.pub
-#expect "Enter passphrase:"
-#send "654321"
 
+
+#\cp -rf /srv/hosts/config/docker_sudoers /etch/sudoers.d
 
 #A few tips to help make your recording successful:
 #
